Update thread jumps on InternalStore and NetworkService (#354)
`finalizeValues` shouldn't call the callback inside the `storeQueue`, as
that would prevent other store operations from happening. This PR adds a
parameter to define the DispatchQueue to use for the callback for
convenience and to potentially avoid an unnecessary thread jump.

`saveValues` doesn't need to use the main thread for its completion
block.